# Attempt to refactor service_instance_list_fetcher perm logic