Rather than passing the list of space ids, pass a dataset that is properly scoped

* A short explanation of the proposed change:
While working on allowing api endpoints to properly return shared routes, we found service instances already have a pattern for this. However they also pass space ids around rather than the datasets as most v3 controllers do.

The space visibility queries are rather complex though, so using it as a subquery concerned us. 

To see if this is an acceptable change, we used the [cf-performance-tests](https://github.com/cloudfoundry/cf-performance-tests) and wrote a quick test to benchmark the service instances list endpoint.

The results are below. In every case, using the dataset is slightly slower. So the choice is now: Do we want performance? or: Do we want consistency?

---
Number of Spaces: 20  
Service Instance Per Space: 500
#### New Code:
```
  Begin Report Entries >>
      individually::as regular user::GET /v3/service_instances
      Name                                 | N | Min     | Median  | Mean    | StdDev | Max    
      =========================================================================================
      GET /v3/service_instances [duration] | 5 | 382.2ms | 397.3ms | 403.1ms | 15.8ms | 429.8ms
  << End Report Entries
```
#### Old Code:
```
  Begin Report Entries >>
      individually::as regular user::GET /v3/service_instances
      Name                                 | N | Min     | Median  | Mean    | StdDev | Max    
      =========================================================================================
      GET /v3/service_instances [duration] | 5 | 375.8ms | 410.3ms | 407.5ms | 25.1ms | 439.6ms
  << End Report Entries
```
---
Number of Spaces: 20000  
Service Instance Per Space: 5
#### New Code:
```
  Begin Report Entries >>
      individually::as regular user::GET /v3/service_instances
      Name                                 | N | Min     | Median  | Mean    | StdDev | Max    
      =========================================================================================
      GET /v3/service_instances [duration] | 5 | 1.4303s | 1.4915s | 1.4852s | 30.3ms | 1.5233s
  << End Report Entries
```
#### Old Code:
```
  Begin Report Entries >>
      individually::as regular user::GET /v3/service_instances
      Name                                 | N | Min     | Median  | Mean    | StdDev | Max    
      =========================================================================================
      GET /v3/service_instances [duration] | 5 | 1.3077s | 1.3323s | 1.3378s | 21.4ms | 1.3664s
  << End Report Entries
```
---
Number of Spaces: 200  
Service Instance Per Space: 5000
#### New Code:
```
  Begin Report Entries >>
      individually::as regular user::GET /v3/service_instances
      Name                                 | N | Min    | Median  | Mean    | StdDev | Max     
      =========================================================================================
      GET /v3/service_instances [duration] | 5 | 9.808s | 9.8519s | 9.8814s | 82.6ms | 10.0384s
  << End Report Entries
```
#### Old Code:
```
  Begin Report Entries >>
      individually::as regular user::GET /v3/service_instances
      Name                                 | N | Min     | Median  | Mean    | StdDev | Max    
      =========================================================================================
      GET /v3/service_instances [duration] | 5 | 9.0863s | 9.1712s | 9.1931s | 75ms   | 9.2905s
  << End Report Entries
```